api/requirements.txt
# FastAPI
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
pydantic-settings>=2.0.0
email-validator>=1.1.3

# db
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.1

# auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# store
minio>=7.1.0

# test
pytest>=6.2.5
pytest-asyncio>=0.18.0
httpx>=0.18.2

# filetype
filetype>=1.2.0

# timezone
pytz>=2024.1

# image processing
Pillow>=10.0.0
selenium>=4.0.0  

# excel
xlsxwriter>=3.0.0  
openpyxl>=3.0.0   

# date science
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
xlsxwriter>=3.0.0

# email 
fastapi-mail>=1.2.0

# task schedule
apscheduler>=3.9.0

# data visualization
plotly>=5.5.0
kaleido>=0.2.1  

# machine learning
scikit-learn>=1.0.0 

# report generation
Jinja2>=3.0.0
html2text>=2020.1.16
reportlab>=4.0.0
markdown>=3.3.0 